% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeOldFSIndex.R
\name{make.fs.index}
\alias{make.fs.index}
\alias{makeOldFSIndex}
\title{Makes ...}
\usage{
makeOldFSIndex(dat, do.plot = T, pr.stn = "33")
}
\arguments{
\item{dat}{The data}
}
\description{
\code{make.fs.index} makes ...
}
